refactor(magicLink): add magic_link requests to dedicated file
