+++ src/api.ts
@@ -2,49 +2,11 @@ import { validateApiKey } from "./api/endUserApiKeys"
 import {
     AccessTokenCreationException,
     ApiKeyValidateException,
-    MagicLinkCreationException,
     MigrateUserException,
     UserNotFoundException,
 } from "./exceptions"
 import { httpRequest } from "./http"
 import { OrgApiKeyValidation, OrgMemberInfo, PersonalApiKeyValidation, User } from "./user"
-
-export type CreateMagicLinkRequest = {
-    email: string
-    redirectToUrl?: string
-    expiresInHours?: string
-    createNewUserIfOneDoesntExist?: boolean
-}
-
-export type MagicLink = {
-    url: string
-}
-
-export function createMagicLink(
-    authUrl: URL,
-    integrationApiKey: string,
-    createMagicLinkRequest: CreateMagicLinkRequest
-): Promise<MagicLink> {
-    const request = {
-        email: createMagicLinkRequest.email,
-        redirect_to_url: createMagicLinkRequest.redirectToUrl,
-        expires_in_hours: createMagicLinkRequest.expiresInHours,
-        create_new_user_if_one_doesnt_exist: createMagicLinkRequest.createNewUserIfOneDoesntExist,
-    }
-    return httpRequest(authUrl, integrationApiKey, `/api/backend/v1/magic_link`, "POST", JSON.stringify(request)).then(
-        (httpResponse) => {
-            if (httpResponse.statusCode === 401) {
-                throw new Error("integrationApiKey is incorrect")
-            } else if (httpResponse.statusCode === 400) {
-                throw new MagicLinkCreationException(httpResponse.response)
-            } else if (httpResponse.statusCode && httpResponse.statusCode >= 400) {
-                throw new Error("Unknown error when creating magic link")
-            }
-
-            return JSON.parse(httpResponse.response)
-        }
-    )
-}
 
 export type CreateAccessTokenRequest = {
     userId: string
